runs/kit/REQ-017/requirements.txt
# REQ-017: Campaign dashboard stats API
# Runtime dependencies

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0

# Test dependencies
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
aiosqlite>=0.19.0